Feat : Added Getting Sensor History and Sensor History By Id

+++ VendManager/VendManager.Domain/SensorValueHistory.cs
@@ -6,7 +6,7 @@ namespace VendManager.Domain
     {
 
         public DateTime DateTimeUTC { get; set; }
-        public int SensorID { get; set; }
+        public long SensorID { get; set; }
         public bool HasStock { get; set; }
     }
 }
